✨ refactor: Update session model import paths and enhance session creation logic in AuthService
