+++ api/cache/banViolation.js
@@ -1,6 +1,6 @@
 const { ViolationTypes } = require('librechat-data-provider');
 const { isEnabled, math, removePorts } = require('~/server/utils');
-const { deleteAllUserSessions } = require('~/models/Session');
+const { deleteAllUserSessions } = require('~/models');
 const getLogStores = require('./getLogStores');
 const { logger } = require('~/config');
